fixup! osutil: fix failing mac os unit tests
