+++ osutil/dev_darwin.go
@@ -1,0 +1,35 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+
+/*
+ * Copyright (C) 2025 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package osutil
+
+// Major obtains the major device number.
+func Major(dev uint64) uint64 {
+	panic(ErrDarwin)
+}
+
+// Minor obtains the minor device number.
+func Minor(dev uint64) uint64 {
+	panic(ErrDarwin)
+}
+
+// Makedev constructs device number from major/minor numbers.
+func Makedev(maj, min uint64) uint64 {
+	panic(ErrDarwin)
+}
