+++ osutil/stat.go
@@ -1,7 +1,7 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
 
 /*
- * Copyright (C) 2014-2015 Canonical Ltd
+ * Copyright (C) 2014-2025 Canonical Ltd
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as
@@ -18,9 +18,6 @@
  */
 
 package osutil
-
-// #include <sys/sysmacros.h>
-import "C"
 
 import (
 	"os"
@@ -150,19 +147,4 @@ func ComparePathsByDeviceInode(a, b string) (match bool, err error) {
 	}
 
 	return os.SameFile(fi1, fi2), nil
-}
-
-// Major obtains the major device number.
-func Major(dev uint64) uint64 {
-	return uint64(C.gnu_dev_major((C.ulong)(dev)))
-}
-
-// Minor obtains the minor device number.
-func Minor(dev uint64) uint64 {
-	return uint64(C.gnu_dev_minor((C.ulong)(dev)))
-}
-
-// Makedev constructs device number from major/minor numbers.
-func Makedev(maj, min uint64) uint64 {
-	return uint64(C.gnu_dev_makedev((C.uint)(maj), (C.uint)(min)))
 }
