+++ osutil/stat_test.go
@@ -1,7 +1,7 @@
 // -*- Mode: Go; indent-tabs-mode: t -*-
 
 /*
- * Copyright (C) 2014-2015 Canonical Ltd
+ * Copyright (C) 2014-2025 Canonical Ltd
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as
@@ -361,11 +361,4 @@ func (ts *StatTestSuite) TestComparePathsByDeviceInodeErrorPathNotExist(c *C) {
 	match, err = osutil.ComparePathsByDeviceInode(base, filepath.Join(base, "missing-dir"))
 	c.Assert(err, ErrorMatches, "*: no such file or directory")
 	c.Assert(match, Equals, false)
-}
-
-func (ts *StatTestSuite) TestDeviceMajorMinor(c *C) {
-	dev := osutil.Makedev(11, 12)
-	c.Check(osutil.Major(dev), Equals, uint64(11))
-	c.Check(osutil.Minor(dev), Equals, uint64(12))
-	c.Check(osutil.Makedev(osutil.Major(dev), osutil.Minor(dev)), Equals, dev)
 }
